+++ geniusyield-server-lib/src/GeniusYield/Server/Assets.hs
@@ -4,6 +4,7 @@ module GeniusYield.Server.Assets (
 ) where
 
 import Fmt
+import GHC.TypeLits (AppendSymbol)
 import GeniusYield.OrderBot.Domain.Assets
 import GeniusYield.Server.Ctx
 import GeniusYield.Server.Utils
@@ -11,7 +12,7 @@ import GeniusYield.Types
 import RIO hiding (logDebug, logInfo)
 import Servant
 
-type AssetsAPI = Summary "Get assets information" :> Description "Get information for a specific asset." :> Capture "asset" GYAssetClass :> Get '[JSON] AssetDetails
+type AssetsAPI = Summary "Get assets information" :> Description ("Get information for a specific asset. " `AppendSymbol` CommonMaestroKeyRequirementText) :> Capture "asset" GYAssetClass :> Get '[JSON] AssetDetails
 
 handleAssetsApi ∷ Ctx → GYAssetClass → IO AssetDetails
 handleAssetsApi ctx@Ctx {..} ac = do