+++ geniusyield-dex-api/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for geniusyield-dex-api
 
+## 0.2.1.0 -- 2024-05-07
+
+* Adds `placePartialOrder''`, `placePartialOrderWithVersion''` to also return for order's NFT token.
+* Exports `placePartialOrder''`, `placePartialOrderWithVersion`, `placePartialOrderWithVersion'` and `placePartialOrderWithVersion''`.
+
 ## 0.2.0.0 -- 2024-04-17
 
 * Adds support for v1.1 family of scripts.